Add endpoint to retrieve open duels for a campaign
A new GET endpoint has been added to the ActionsController class to retrieve open duels for a specific campaign. The route for this endpoint is Campaigns/{campaignId}/[controller]/duels, and it returns a list of DuelAction objects.

In ActionsService.cs, a new method DuelsOpen has been added to retrieve a list of open duel rooms for a given campaign and user, mapping them to DuelAction objects. The IActionsService interface has been updated to include this new method.

Additionally, a using directive for System.Collections.Generic has been added to ActionsService.cs.